/Matchups, /matchups/id, /teams, /teams/id REST controllers were implemented

+++ src/main/java/com/aryanrt/stats/repositories/MatchupRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import com.aryanrt.stats.models.Matchup;
 
 @RepositoryRestResource(collectionResourceRel = "matchup", path = "matchup")
-public interface MatchupRepository extends PagingAndSortingRepository<Matchup, Long> {
+public interface MatchupRepository extends PagingAndSortingRepository<Matchup, Integer> {
 
-    Matchup findById(long id);
+    Matchup findById(int id);
 }
